chore: Move MockedProvider to provider-test-tools
