+++ packages/provider-test-tools/src/index.js
@@ -1,6 +1,7 @@
 import clientTester from "./client-tester";
 import providerTester from "./provider-tester";
 import fixtureGenerator from "./fixture-generator";
+import MockedProvider from "./mocked-provider";
 
 export * from "./helpers";
-export { clientTester, providerTester, fixtureGenerator };
+export { clientTester, fixtureGenerator, MockedProvider, providerTester };
