+++ packages/provider/__tests__/author-articles-with-images.test.js
@@ -1,7 +1,9 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import { MockedProvider } from "@times-components/utils";
-import { fixtureGenerator } from "@times-components/provider-test-tools";
+import {
+  fixtureGenerator,
+  MockedProvider
+} from "@times-components/provider-test-tools";
 import { AuthorArticlesWithImagesProvider } from "../src/provider";
 
 const renderComponent = child =>
