+++ packages/provider/provider.showcase.js
@@ -5,7 +5,7 @@ import topicFixture from "@times-components/provider-test-tools/fixtures/topic.j
 import articleFixture from "@times-components/provider-test-tools/fixtures/article.json";
 import { addTypenameToDocument } from "apollo-utilities";
 import gql from "graphql-tag";
-import { MockedProvider } from "@times-components/utils";
+import { MockedProvider } from "@times-components/provider-test-tools";
 import { fixtureGenerator } from "@times-components/provider-test-tools";
 import connectGraphql, {
   ArticleProvider,
