+++ packages/topic/__tests__/topic-functional.js
@@ -1,8 +1,11 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import mockDate from "mockdate";
-import { fixtureGenerator } from "@times-components/provider-test-tools";
-import { delay, MockedProvider } from "@times-components/utils";
+import {
+  fixtureGenerator,
+  MockedProvider
+} from "@times-components/provider-test-tools";
+import { delay } from "@times-components/utils";
 import Topic from "../src/topic";
 
 // This is the only possible way for this to work... :'-(
